+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [0.2.3] - 2024-07-13
 ### Added
 - Scientific notation formatting for arbitrary columns when generating Latex
   tables in the `Table` class.
@@ -81,7 +82,8 @@ Downstream moderate risk update release.
 
 
 <!-- links -->
-[Unreleased]: https://github.com/plandes/datdesc/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/plandes/datdesc/compare/v0.2.3...HEAD
+[0.2.3]: https://github.com/plandes/datdesc/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/plandes/datdesc/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/plandes/datdesc/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/plandes/datdesc/compare/v0.1.1...v0.2.0